+++ doc/changelog.rst
@@ -4,6 +4,8 @@ v0.3.8 (*unreleased*)
 ---------------------
 - use the `doctest` formatter for doctest lines in `rst` code blocks (:issue:`150`, :pull:`151`)
 - drop support for ``python=3.6`` (:pull:`153`)
+- replace the custom color formatting code with `rich <https://github.com/textualize/rich>`_
+  (:issue:`146`, :pull:`157`).
 
 v0.3.7 (13 September 2022)
 --------------------------